+++ .gitignore
@@ -1,17 +1,5 @@
 **/*.rs.bk
-/CHANGELOG.md
 /book/book
-/book/src/SUMMARY.md
-/book/src/bittorrent.md
-/book/src/changelog.md
-/book/src/commands.md
-/book/src/commands/
-/book/src/faq.md
-/book/src/introduction.md
-/book/src/references
-/book/src/references.md
-/completions
-/man
 /target
 /wiki
 /www/book
