+++ bin/gen/Cargo.toml
@@ -11,10 +11,9 @@ askama            = "0.9.0"
 cargo_toml        = "0.8.0"
 chrono            = "0.4.11"
 fehler            = "1.0.0"
+fs_extra          = "1.1.0"
 git2              = "0.13.1"
 globset           = "0.4.5"
-ignore            = "0.4.14"
-lexiclean         = "0.0.1"
 libc              = "0.2.69"
 log               = "0.4.8"
 pretty_env_logger = "0.4.0"
@@ -25,7 +24,6 @@ structopt         = "0.3.12"
 strum             = "0.18.0"
 strum_macros      = "0.18.0"
 tempfile          = "3.1.0"
-walkdir           = "2.3.1"
 
 [dependencies.serde]
 version  = "1.0.106"
