+++ bin/gen/src/common.rs
@@ -17,8 +17,6 @@ pub(crate) use cargo_toml::Manifest;
 pub(crate) use chrono::{DateTime, NaiveDateTime, Utc};
 pub(crate) use fehler::{throw, throws};
 pub(crate) use git2::{Commit, Oid, Repository};
-pub(crate) use ignore::overrides::OverrideBuilder;
-pub(crate) use lexiclean::Lexiclean;
 pub(crate) use libc::EXIT_FAILURE;
 pub(crate) use log::info;
 pub(crate) use regex::Regex;
@@ -29,7 +27,6 @@ pub(crate) use structopt::StructOpt;
 pub(crate) use strum::VariantNames;
 pub(crate) use strum_macros::{EnumVariantNames, IntoStaticStr};
 pub(crate) use url::Url;
-pub(crate) use walkdir::WalkDir;
 
 // modules
 pub(crate) use crate::error;
@@ -39,9 +36,9 @@ pub(crate) use crate::{command_ext::CommandExt, row::Row, slug::Slug, template_e
 
 // structs and enums
 pub(crate) use crate::{
-  bin::Bin, changelog::Changelog, config::Config, entry::Entry, error::Error, example::Example,
-  faq::Faq, faq_entry::FaqEntry, introduction::Introduction, kind::Kind, metadata::Metadata,
-  opt::Opt, package::Package, project::Project, readme::Readme, reference::Reference,
-  reference_section::ReferenceSection, release::Release, subcommand::Subcommand, summary::Summary,
-  table::Table,
+  arguments::Arguments, bin::Bin, bin_subcommand::BinSubcommand, changelog::Changelog,
+  config::Config, entry::Entry, error::Error, example::Example, faq::Faq, faq_entry::FaqEntry,
+  introduction::Introduction, kind::Kind, metadata::Metadata, options::Options, package::Package,
+  project::Project, readme::Readme, reference::Reference, reference_section::ReferenceSection,
+  release::Release, subcommand::Subcommand, summary::Summary, table::Table,
 };
