+++ bin/gen/src/main.rs
@@ -3,7 +3,9 @@ use crate::common::*;
 #[macro_use]
 mod cmd;
 
+mod arguments;
 mod bin;
+mod bin_subcommand;
 mod changelog;
 mod command_ext;
 mod common;
@@ -16,7 +18,7 @@ mod faq_entry;
 mod introduction;
 mod kind;
 mod metadata;
-mod opt;
+mod options;
 mod package;
 mod project;
 mod readme;
@@ -33,7 +35,7 @@ mod template_ext;
 fn main() {
   pretty_env_logger::init();
 
-  if let Err(error) = Opt::from_args().run() {
+  if let Err(error) = Arguments::from_args().run() {
     let bold = Style::new().bold();
     let red = Style::new().fg(ansi_term::Color::Red).bold();
     eprintln!("{}: {}", red.paint("error"), bold.paint(error.to_string()));
