+++ book/src/SUMMARY.md
@@ -1,0 +1,1 @@
+../../target/gen/book/SUMMARY.md
