+++ book/src/bittorrent.md
@@ -1,0 +1,1 @@
+../../target/gen/book/bittorrent.md
