+++ book/src/changelog.md
@@ -1,0 +1,1 @@
+../../target/gen/book/changelog.md
