+++ book/src/commands
@@ -1,0 +1,1 @@
+../../target/gen/book/commands
