+++ book/src/commands.md
@@ -1,0 +1,1 @@
+../../target/gen/book/commands.md
