+++ book/src/faq.md
@@ -1,0 +1,1 @@
+../../target/gen/book/faq.md
