+++ book/src/introduction.md
@@ -1,0 +1,1 @@
+../../target/gen/book/introduction.md
