+++ book/src/references
@@ -1,0 +1,1 @@
+../../target/gen/book/references
