+++ book/src/references.md
@@ -1,0 +1,1 @@
+../../target/gen/book/references.md
